+++ CHANGES.md
@@ -1,5 +1,10 @@
 
+* **FEATURE:** recipe & command-line tool for master dark creation (`mdark`)
+* **ENHANCEMENT:** default binning add to detector settings file
+* **ENHANCEMENT:** added mixed exposure time unit test for dark-frames
 * **ENHANCEMENT:** added default values for gain and ron in the detector settings files. Default values can be overwritten if correct GAIN and RON are found in fits-headers (overwritten for UVB and VIS but not NIR for XShooter)
+* **FIXED:** binning factor used when trimming frames
+* **FIXED:** the trimming dimensions of NIR frames - bad-pixel map now aligns correctly with data frame
 * **FIXED:** science pixels for all 3 xshooter detectors in parameters file
 
 ## v0.3.0 - August 18, 2020